# web/lib/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomFoodCreate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 255,
            "title": "Name",
            "type": "string"
          },
          "nutrients": {
            "$ref": "#/components/schemas/NutrientProfile"
          },
          "portions": {
            "items": {
              "$ref": "#/components/schemas/HouseholdPortion-Input"
            },
            "maxItems": 20,
            "title": "Portions",
            "type": "array"
          }
        },
        "required": [
          "name",
          "nutrients"
        ],
        "title": "CustomFoodCreate",
        "type": "object"
      },
...
      "FoodCapabilities": {
        "properties": {
          "barcode_lookup_enabled": {
            "title": "Barcode Lookup Enabled",
            "type": "boolean"
          },
          "schema_version": {
            "const": "1.0",
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "barcode_lookup_enabled"
        ],
        "title": "FoodCapabilities",
        "type": "object"
      },
...
      "OpenFoodFactsExport": {
        "properties": {
          "contents_license": {
            "const": "DbCL-1.0",
            "default": "DbCL-1.0",
            "title": "Contents License",
            "type": "string"
          },
          "contents_license_url": {
            "default": "https://opendatacommons.org/licenses/dbcl/1-0/",
            "title": "Contents License Url",
            "type": "string"
          },
          "database_license": {
            "const": "ODbL-1.0",
            "default": "ODbL-1.0",
            "title": "Database License",
            "type": "string"
          },
          "database_license_url": {
            "default": "https://opendatacommons.org/licenses/odbl/1-0/",
            "title": "Database License Url",
            "type": "string"
          },
          "dataset": {
            "default": "opennosh-open-food-facts-cache",
            "title": "Dataset",
            "type": "string"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/OpenFoodFactsExportEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "notice": {
            "default": "Contains information from Open Food Facts, made available under the Open Database License. Individual contents are available under the Database Contents License.",
            "title": "Notice",
            "type": "string"
          },
          "schema_version": {
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "source": {
            "default": "Open Food Facts",
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "default": "https://world.openfoodfacts.org/",
            "title": "Source Url",
            "type": "string"
          }
        },
        "required": [
          "entries"
        ],
        "title": "OpenFoodFactsExport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "opennosh API",
    "version": "0.26.0.0",
    "x-opennosh-contract-version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/export/foods/odbl": {
      "get": {
        "operationId": "open_food_facts_export_api_v1_export_foods_odbl_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenFoodFactsExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The current user is not authorized."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with the latest state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request failed validation."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The server could not complete the request."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service returned an unusable response."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The service is temporarily unavailable."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service timed out."
          }
        },
        "summary": "Open Food Facts Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/export/foods/openfoodfacts": {
      "get": {
        "operationId": "open_food_facts_export_api_v1_export_foods_openfoodfacts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenFoodFactsExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The current user is not authorized."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with the latest state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request failed validation."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The server could not complete the request."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service returned an unusable response."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The service is temporarily unavailable."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service timed out."
          }
        },
        "summary": "Open Food Facts Export",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/v1/foods/barcode/{barcode}": {
      "get": {
        "operationId": "barcode_lookup_api_v1_foods_barcode__barcode__get",
        "parameters": [
          {
            "in": "path",
            "name": "barcode",
            "required": true,
            "schema": {
              "maxLength": 14,
              "minLength": 8,
              "title": "Barcode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpenFoodFactsFood"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The current user is not authorized."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with the latest state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request failed validation."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The server could not complete the request."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service returned an unusable response."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The service is temporarily unavailable."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service timed out."
          }
        },
        "summary": "Barcode Lookup",
        "tags": [
          "foods"
        ]
      }
    },
    "/api/v1/foods/capabilities": {
      "get": {
        "operationId": "capabilities_api_v1_foods_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodCapabilities"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The current user is not authorized."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with the latest state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request failed validation."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The server could not complete the request."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service returned an unusable response."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The service is temporarily unavailable."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service timed out."
          }
        },
        "summary": "Capabilities",
        "tags": [
          "foods"
        ]
      }
    },
    "/api/v1/foods/custom": {
      "post": {
        "operationId": "custom_food_create_api_v1_foods_custom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFoodCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFoodResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request is invalid."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication is required."
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The current user is not authorized."
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The requested resource was not found."
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request conflicts with the latest state."
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request failed validation."
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The request rate limit was exceeded."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The server could not complete the request."
          },
          "502": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service returned an unusable response."
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "The service is temporarily unavailable."
          },
          "504": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "An upstream service timed out."
          }
        },
        "summary": "Custom Food Create",
        "tags": [
          "foods"
        ]
      }
    },
...
  }
}